Add files via upload
Check the code rearrangement and see how to arrange your code lines. Especially in the css.

Also, why is your nav in your head tags and not in the body tags? Is there a reason for that? Tell me?
There are key components missing in your navigation menu.

https://www.w3schools.com/css/css_navbar.asp

Use the link I have sent above and compare what you have done carefully with the tutorial.

Make corrections and revert back to me.